+++ run.sh
@@ -13,4 +13,4 @@ pip install llama-index-vector-stores-lancedb
 pip install llama-index-embeddings-clip
 pip install git+https://github.com/openai/CLIP.git
 export LD_PRELOAD=/usr/local/lib/python3.8/dist-packages/scikit_learn.libs/libgomp-d22c30c5.so.1.0.0
-python3 ./mulitRAG_old.py
+python3 ./mulitRAG.py
